use BaseLayoutHeader component #1213
